+++ src/backend/xray/__init__.py
@@ -1,3 +1,3 @@
-from xray.xray import create_lifespan
+from backend.xray.main import create_lifespan
 
 __all__ = ["create_lifespan"]
